chore: move all additional desktop / mobile components into parent dirs
