+++ apps/marginfi-v2-ui/src/components/desktop/Banner/index.tsx
@@ -1,0 +1,1 @@
+export * from "./Banner";
