+++ apps/marginfi-v2-ui/src/components/desktop/CampaignWizard/index.tsx
@@ -1,0 +1,1 @@
+export * from "./CampaignWizard";
