+++ apps/marginfi-v2-ui/src/components/mobile/EmissionsBanner/index.tsx
@@ -1,0 +1,1 @@
+export * from "./EmissionsBanner";
